Updating to newest code
Updating to newest code. This code does crash due to some issues with the road shapefiles from the municipality in the greenspace generation function. However,  it is the most up to date and tends to work if the geometry is not abnormal.
The comments are not fully up to date however and might be a bit lacking